+++ src/rogallo/widgets/viewer/title.py
@@ -9,7 +9,12 @@ from textual.reactive import var
 from textual.widgets import Label
 
 ##############################################################################
+# Wasat imports.
+from wasat import VerificationMethod
+
+##############################################################################
 # Local imports.
+from ...data import load_configuration
 from ...types import RogalloLocation
 
 
@@ -24,9 +29,19 @@ class ViewerTitle(Horizontal):
         height: 1;
         padding: 0 1;
 
-        #lock-icon {
-            width: 1;
+        #verification-method, #lock-icon {
+            width: 2;
+            padding-right: 1;
+        }
+
+        .--verified-ca, #lock-icon {
             color: $text-success;
+        }
+        .--verified-tofu {
+            color: $text-warning;
+        }
+        .--verified-off, .--verified-none {
+            color: $text-error;
         }
 
         #location {
@@ -38,9 +53,13 @@ class ViewerTitle(Horizontal):
 
     location: var[RogalloLocation | None] = var(None, always_update=True)
     """The location to display."""
+    verification_method: var[VerificationMethod | None] = var(None)
+    """The verification method used for the connection."""
     needed_certificate: var[bool] = var(False)
     """Whether the location needed a certificate."""
 
+    _verification_method_icon = query_one("#verification-method", Label)
+    """The label for the verification method."""
     _lock_icon = query_one("#lock-icon", Label)
     """The label for the lock icon."""
     _location_label = query_one("#location", Label)
@@ -48,6 +67,7 @@ class ViewerTitle(Horizontal):
 
     def compose(self) -> ComposeResult:
         """Compose the child widgets."""
+        yield Label(id="verification-method")
         yield Label(id="lock-icon")
         yield Label(id="location")
 
@@ -66,7 +86,34 @@ class ViewerTitle(Horizontal):
 
     def _watch_needed_certificate(self) -> None:
         """React to the needed_certificate changing."""
-        self._lock_icon.update("\u26bf" if self.needed_certificate else " ")
+        self._lock_icon.update(
+            load_configuration().client_certificate_used_icon
+            if self.needed_certificate
+            else " "
+        )
+
+    def _watch_verification_method(self) -> None:
+        """React to the verification_method changing."""
+        self._verification_method_icon.set_classes(
+            f"--verified-{str(self.verification_method).lower()}"
+        )
+        match self.verification_method:
+            case "ca":
+                self._verification_method_icon.update(
+                    load_configuration().verified_ca_icon
+                )
+            case "tofu":
+                self._verification_method_icon.update(
+                    load_configuration().verified_tofu_icon
+                )
+            case "off":
+                self._verification_method_icon.update(
+                    load_configuration().verified_off_icon
+                )
+            case _:
+                self._verification_method_icon.update(
+                    load_configuration().verified_none_icon
+                )
 
     def on_resize(self) -> None:
         """Handle the widget being resized."""